fix: use strategy method for registering crafter&client, adding GlobalExceptionHandler and refactoring the code

+++ Backend/CommuniCraft/src/main/java/com/communicate_craft/service/LocationService.java
@@ -6,6 +6,6 @@ import java.util.Optional;
 
 public interface LocationService {
     Location saveLocation(Location location);
-    Optional<Location> findById(Integer locationId);
+    Optional<Location> findById(Long locationId);
     Optional<Location> checkIfLocationExists(String cityName, String stateName, String countryName);
 }
